+++ src/Hyyan/WPI/Reports.php
@@ -32,6 +32,20 @@ class Reports
     protected $report;
 
     /**
+     * Combine button styling
+     *
+     * @var array
+     */
+    protected $combineStyle = array(
+        'height:inherit',
+        'float:right',
+        'line-height:26px',
+        'padding: 10px',
+        'display: block',
+        'text-decoration: none'
+    );
+
+    /**
      * Construct object
      */
     public function __construct()
@@ -58,6 +72,9 @@ class Reports
                 'woocommerce_report_low_in_stock_query_from'
                 , array($this, 'filterStockByLangauge')
         );
+
+        /* Combine product report with its translation */
+        add_action('admin_init', array($this, 'translateProductIDS'));
     }
 
     /**
@@ -81,6 +98,11 @@ class Reports
             'sales_by_category'
         );
         if (!in_array($this->report, $reports)) {
+            return $query;
+        }
+
+        /* Check for product_ids */
+        if (isset($_GET['product_ids'])) {
             return $query;
         }
 
@@ -116,6 +138,105 @@ class Reports
         $where = $polylang->model->where_clause($lang, 'post');
 
         return str_replace('WHERE 1=1', "{$join} WHERE 1=1 {$where}", $query);
+    }
+
+    /**
+     * Translate product IDS for product report
+     *
+     * @global \Polylang $polylang
+     * @global \WooCommerce $woocommerce
+     *
+     * @return false if woocommerce or polylang not found
+     */
+    public function translateProductIDS()
+    {
+        global $polylang, $woocommerce;
+        if (!$polylang || !$woocommerce) {
+            return false;
+        }
+
+        /* Check for product_ids */
+        if (!isset($_GET['product_ids'])) {
+            return false;
+        }
+
+        /* Show combine button anyway */
+        add_action(
+                'admin_print_scripts'
+                , array($this, 'showCombineButton')
+                , 100
+        );
+
+        $IDS = (array) $_GET['product_ids'];
+        $extendedIDS = array();
+
+        if (static::isCombine()) {
+
+            foreach ($IDS as $ID) {
+                $translations = Utilities::getProductTranslationsArrayByID($ID);
+                $extendedIDS = array_merge($extendedIDS, $translations);
+            }
+        } elseif (isset($_GET['lang'])) {
+
+            $lang = esc_attr($_GET['lang']);
+            foreach ($IDS as $ID) {
+                $translation = Utilities::getProductTranslationByID($ID, $lang);
+                $extendedIDS[] = $translation->id;
+            }
+        }
+
+        /* Update with extended list */
+        if (!empty($extendedIDS)) {
+            $_GET['product_ids'] = $extendedIDS;
+        }
+    }
+
+    /**
+     * Show the combine buttton for product report
+     */
+    public function showCombineButton()
+    {
+        $isCombine = static::isCombine();
+        $url = Utilities::getCurrentAdminUrl();
+
+        if (!$isCombine) {
+            $url = add_query_arg(array('combine' => '', 'lang' => 'all'), $url);
+        } else {
+            $url = remove_query_arg('combine', $url);
+        }
+
+        $title = __('Combine Report', 'woo-poly-integration');
+        $desc = __('Combine Report With Its Translation', 'woo-poly-integration');
+        $icon = $isCombine ?
+                '<span class=\"dashicons dashicons-yes\" style=\"line-height:inherit\"></span>' :
+                '<span class=\"dashicons dashicons-editor-contract\" style=\"line-height:inherit\"></span>';
+        $style = $isCombine ? implode(';', $this->combineStyle) . ';color:green' :
+                implode(';', $this->combineStyle);
+
+        printf(
+                '<script type="text/javascript" id="woo-poly-combine-report">'
+                . ' jQuery(document).ready(function ($) {'
+                . '     $(".stats_range").prepend("<a href=\'%s\' style=\'%s\' title=\'%s\'>%s</a>")'
+                . ' });'
+                . '</script>'
+                , $url
+                , $style
+                , $desc
+                , $icon . $title
+        );
+    }
+
+    /**
+     * Is combine
+     *
+     * Check if combine mode is requested
+     *
+     * @return boolean true if combine mode , false otherwise
+     */
+    public static function isCombine()
+    {
+        return isset($_GET['combine']) &&
+                (isset($_GET['lang']) && esc_attr($_GET['lang']) === 'all');
     }
 
 }